Move `CheerioAPI` definition to load.ts
